requirements.txt
jupyter
matplotlib
numpy>=1.21
pandas~=1.2.3
scipy~=1.6.1
seaborn
scikit-learn
statsmodels
pytz~=2021.1
python-dateutil~=2.8.1
openpyxl
biopython
